+++ postgres/src/driver/generic.rs
@@ -367,8 +367,13 @@ impl DriverRx {
                 ResponseMessage::Normal(mut msg) => {
                     // lock the shared state only when needed and keep the lock around a bit for possible multiple messages
                     let inner = guard.get_or_insert_with(|| self.guarded.lock().unwrap());
-                    let complete = msg.complete();
-                    let _ = inner.res.front_mut().ok_or_else(|| msg.parse_error())?.send(msg);
+                    let complete = msg.complete;
+
+                    if !complete {
+                        println!("message is not finished");
+                    }
+
+                    let _ = inner.res.front_mut().ok_or_else(|| msg.parse_error())?.send(msg.buf);
                     if complete {
                         inner.res.pop_front();
                     }